Update starting branch: Rename BuenaToken to BuenoToken for mainnet
- Rename contract from BuenaToken to BuenoToken (matching deployed contract)
- Update contract address to mainnet: 0xCFA45ECA955dd195b5b5Fc0E40d1A1B06f16793C
- Update start block to 50636395 (mainnet deployment block)
- Change network from celo-sepolia to celo (mainnet)
- Rename all artifact files to match BuenoToken naming
- Update subgraph configuration for mainnet deployment
- Update all imports and references throughout codebase
- Keep Ownable-based events (students will migrate to AccessControl)

This is the STARTING point for the workshop. Students will:
1. Understand the deployed contract structure
2. Learn to migrate from Ownable to AccessControl
3. Update subgraph schema and handlers
4. Implement data fetching in the frontend
