+++ packages/subgraph/networks.json
@@ -1,8 +1,8 @@
 {
-  "celo-sepolia": {
-    "BuenaToken": {
-      "address": "0xa0304b26DEb6C0D12416F4bDbB09e52e6f800A1F",
-      "startBlock": 7209007
+  "celo": {
+    "BuenoToken": {
+      "address": "0xCFA45ECA955dd195b5b5Fc0E40d1A1B06f16793C",
+      "startBlock": 50636395
     }
   }
 }
